class path and application name renamed
