+++ ehcache/pom.xml
@@ -62,15 +62,14 @@
 	</dependencies>
 
 	<build>
-
 		<resources>
 			<resource>
 				<directory>${basedir}/src/main/resources</directory>
 				<filtering>true</filtering>
 				<includes>
 					<include>**/application*.properties</include>
+					<include>**/application.properties</include>
 					<include>**/banner.txt</include>
-					<include>**/ehcache.xml</include>
 				</includes>
 			</resource>
 		</resources>
@@ -80,6 +79,9 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 				<version>${spring-boot-dependencies.version}</version>
+				<configuration>
+					<mainClass>com.c8y.ms.templates.ehcache.App</mainClass>
+				</configuration>
 				<executions>
 					<execution>
 						<goals>
@@ -87,11 +89,7 @@
 						</goals>
 					</execution>
 				</executions>
-				<configuration>
-					<mainClass>${main.class}</mainClass>
-				</configuration>
 			</plugin>
-
 			<plugin>
 				<groupId>com.nsn.cumulocity.clients-java</groupId>
 				<artifactId>microservice-package-maven-plugin</artifactId>
@@ -104,15 +102,14 @@
 							<goal>package</goal>
 						</goals>
 						<configuration>
-							<name>${project.artifactId}</name>
+							<name>templates-ehcache</name>
+							<image>templates-ehcache</image>
 							<encoding>UTF-8</encoding>
-							<rpmSkip>true</rpmSkip>
-							<containerSkip>${c8y.docker.skip}</containerSkip>
+							<skip>${c8y.docker.skip}</skip>
 						</configuration>
 					</execution>
 				</executions>
 			</plugin>
-
 		</plugins>
 	</build>
 </project>
